Move DataType/Entry into mod data
