+++ src/lib.rs
@@ -15,8 +15,8 @@ mod macros;
 pub mod tag;
 pub mod val;
 
+pub use data::{DataType, Entry};
 pub use decode::{Decoded, Decoder};
-pub use dir::DataType;
 pub use error::{
     DecodeError, DecodeErrorKind, DecodeResult, DecodeValueError, FileHeaderError, TagError,
 };
